Added tabs to the CF prowindow. Created view model and views for Edit and Delete CF templates
