+++ VSProjects/Geomapmaker/ViewModels/ContactsFaults/DeleteContactFault.cs
@@ -1,0 +1,12 @@
+﻿namespace Geomapmaker.ViewModels.ContactsFaults
+{
+    public class DeleteContactFault
+    {
+        public ContactsFaultsViewModel ParentVM { get; set; }
+
+        public DeleteContactFault(ContactsFaultsViewModel parentVM)
+        {
+            ParentVM = parentVM;
+        }
+    }
+}
