+++ VSProjects/Geomapmaker/ViewModels/ContactsFaults/EditContactFaultVM.cs
@@ -1,0 +1,13 @@
+﻿namespace Geomapmaker.ViewModels.ContactsFaults
+{
+    public class EditContactFaultVM
+    {
+        public ContactsFaultsViewModel ParentVM { get; set; }
+
+        public EditContactFaultVM(ContactsFaultsViewModel parentVM)
+        {
+            ParentVM = parentVM;
+        }
+
+    }
+}
