+++ VSProjects/Geomapmaker/Views/ContactsFaults/CreateContactFault.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Geomapmaker.Views.ContactsFaults
+{
+    /// <summary>
+    /// Interaction logic for CreateContactFault.xaml
+    /// </summary>
+    public partial class CreateContactFault : UserControl
+    {
+        public CreateContactFault()
+        {
+            InitializeComponent();
+        }
+    }
+}
