+++ VSProjects/Geomapmaker/Views/ContactsFaults/EditContactFault.xaml
@@ -1,0 +1,12 @@
+﻿<UserControl x:Class="Geomapmaker.Views.ContactsFaults.EditContactFault"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Geomapmaker.Views.ContactsFaults"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid>
+        <TextBlock>Edit Contact Fault</TextBlock>
+    </Grid>
+</UserControl>
